--- ROS_Packages/ros_tcp_endpoint/package.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>ros_tcp_endpoint</name>
-  <version>0.3.0</version>
-  <description>Acts as the bridge between Unity messages sent via Websocket and ROS messages.</description>
-
-  <maintainer email="unity-robotics@unity3d.com">Unity Robotics</maintainer>
-
-  <license>Apache 2.0</license>
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>rospy</build_depend>
-  <build_depend>message_generation</build_depend>
-  <build_export_depend>rospy</build_export_depend>
-  <exec_depend>rospy</exec_depend>
-  <exec_depend>message_runtime</exec_depend>
-
-
-  <!-- The export tag contains other, unspecified, tags -->
-  <export>
-    <!-- Other tools can request additional information be placed here -->
-
-  </export>
-</package>
