+++ test/test_publisher.py
