+++ test/test_thread_pauser.py
